removing sections of the original text
creating an abridged version of the adventure that can just be a standalone carnival experience, which can be dropped into any campaign (for ex: Lost Mine of Phandelver, in my case).  Therefore, I'm removing sections of the adventure text that are not needed when not running the full adventure into the Feywild.