removing witchlight hand (small), adding Kester statblock to creatures.json
okay, so it was the witchlight hand (small) that broke the bestiary. i think it might be because i was trying to set it to _copy from a homebrew source (from itself), and yeah okay the data wasn't completely unchnaged from original source now that i think about it -- i did change the 'source' values. So this test determines whether the simplest _copy on the Master Thief statblock is working.